# specs/001-zero-training-mvp/contracts/api-openapi.yaml
openapi: 3.1.0
info:
  title: LiverRa v1 MVP HTTP API
  version: 1.0.0
  description: |
    Public HTTP API for the LiverRa zero-training liver AI pipeline platform.

    **Auth**: OIDC Bearer token from AWS Cognito (research A.1) with `custom:tenant_id` claim.
    **RBAC**: Every protected op declares `x-required-permission`. Step-up auth (≤5 min freshness) is marked `x-step-up: true`.
    **Errors**: RFC 7807 Problem Details. `type`, `title`, `status`, `detail`, plus LiverRa extensions `tenant_id`, `audit_event_id`, `claim_key` (never PHI).
    **Audit**: Every state-changing op emits a FHIR AuditEvent (chain-of-hashes research A.3).
servers:
  - url: https://app.liverra.ai/api/v1
    description: Production
  - url: http://localhost:7050/api/v1
    description: Local dev (FastAPI ml-inference)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Report:
      type: object
      properties:
        id: { type: string, format: uuid }
        analysis_id: { type: string, format: uuid }
        status: { type: string, enum: [draft, finalized, superseded, retracted] }
        finalized_at: { type: string, format: date-time, nullable: true }
        pdf_url: { type: string, format: uri }
        seg_url: { type: string, format: uri }
        sr_url: { type: string, format: uri }
        seg_sop_instance_uid: { type: string }
        sr_sop_instance_uid: { type: string }
        ruo_watermark_present: { type: boolean, const: true }
        superseded_by_report_id: { type: string, format: uuid, nullable: true }
    ReportDelivery:
      type: object
      properties:
        id: { type: string, format: uuid }
        report_id: { type: string, format: uuid }
        artifact_type: { type: string, enum: [seg, sr] }
        destination_ae_title: { type: string }
        status: { type: string, enum: [pending, sending, acknowledged, failed, manual_fallback] }
        retry_count: { type: integer }
        last_error: { type: string, nullable: true, description: "PHI-scrubbed" }
        next_attempt_at: { type: string, format: date-time, nullable: true }
...
  responses:
...
    StepUpRequired:
      description: 401 with WWW-Authenticate per RFC 9470
      headers:
        WWW-Authenticate:
          schema: { type: string }
paths:
...
  /reviews/{review_id}/finalize:
    parameters:
      - { name: review_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [review, export]
      x-required-permission: report.finalize
      x-step-up: true
      summary: Finalize the review → triggers Report generation (US5)
      responses:
        '202':
          description: Report generation queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id: { type: string, format: uuid }
                  status: { type: string, enum: [draft, finalizing] }
        '401':
          $ref: '#/components/responses/StepUpRequired'
  /reports/{report_id}:
    parameters:
      - { name: report_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [export]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
  /reports/{report_id}/pacs-push:
    parameters:
      - { name: report_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [export]
      x-required-permission: report.pacs_push
      summary: Push SEG + SR to hospital PACS (FR-026, FR-026a)
      responses:
        '202':
          description: Delivery attempts started
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ReportDelivery' }
...
  /reports/{report_id}/retract:
    parameters:
      - { name: report_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [export]
      x-required-permission: report.retract
      x-step-up: true
      summary: Retract a finalized Report (FR-027a)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200': { description: Retracted }
...